+++ server/utils/markdown.utils.js
@@ -1,0 +1,4 @@
+const marked = require("marked")
+const parseMarkedDownFile = (markDown) => {
+    return marked.parsed(markDown)
+}